+++ label_bot/commands.py
@@ -20,7 +20,7 @@ EVENT_MAP = {
 RE_COMMANDS = re.compile(
     r'''(?x)
     [ ]+(?:
-        (?P<lgtm>lgtm(?P<lgtm_key>[ ]*[a-z][-a-z0-9_]*(?:[ ]*,[ ]*[a-z][-a-z0-9_]*)*)?) |
+        (?P<lgtm>lgtm) |
         (?P<retrigger>retrigger[ ]+(?P<retrigger_task>auto-labels|wip|review|triage|all)) |
         (?P<sync>sync[ ]+labels)
     )\b
@@ -29,7 +29,7 @@ RE_COMMANDS = re.compile(
 )
 
 
-class Command(namedtuple('Command', ['command', 'event', 'pending', 'live', 'kwargs'])):
+class Command(namedtuple('Command', ['command', 'event', 'pending', 'live'])):
     """Command."""
 
 
@@ -52,15 +52,15 @@ async def command_retrigger(event, action, gh):
     payload[key] = issue
 
     if action in ('triage', 'all') and key == 'issue':
-        command = Command(triage_labels.run, util.Event(event_type, payload), None, False, {})
+        command = Command(triage_labels.run, util.Event(event_type, payload), None, False)
     elif action == 'all' and key == 'pull_request':
-        command = Command(run_all_pull_actions, util.Event(event_type, payload), None, False, {})
+        command = Command(run_all_pull_actions, util.Event(event_type, payload), None, False)
     elif action == 'review' and key == 'pull_request':
-        command = Command(review_labels.run, util.Event(event_type, payload), None, True, {})
+        command = Command(review_labels.run, util.Event(event_type, payload), None, False)
     elif action == 'wip' and key == 'pull_request':
-        command = Command(wip_labels.run, util.Event(event_type, payload), None, True, {})
+        command = Command(wip_labels.run, util.Event(event_type, payload), None, True)
     elif action == 'auto-labels' and key == 'pull_request':
-        command = Command(wildcard_labels.run, util.Event(event_type, payload), wildcard_labels.pending, True, {})
+        command = Command(wildcard_labels.run, util.Event(event_type, payload), wildcard_labels.pending, True)
     else:
         command = None
     return command
@@ -76,15 +76,11 @@ async def command_sync(event, gh):
     return Command(sync_labels.run, util.Event(event_type, payload), sync_labels.pending, False)
 
 
-async def command_lgtm(event, key, gh):
+async def command_lgtm(event, gh):
     """Handle "looks good to me" command."""
 
     if event.data['issue']['state'] != 'open':
         return None
-
-    kwargs = {}
-    if key:
-        kwargs['keys'] = [item.strip() for item in key.split(',')]
 
     await asyncio.sleep(1)
     payload = {'repository': event.data['repository']}
@@ -96,7 +92,7 @@ async def command_lgtm(event, key, gh):
         key = 'pull_request'
         issue = await gh.getitem(issue['pull_request']['url'])
     payload[key] = issue
-    return Command(lgtm_labels.run, util.Event(event_type, payload), None, False, kwargs)
+    return Command(lgtm_labels.run, util.Event(event_type, payload), None, False)
 
 
 async def run_all_pull_actions(event, gh, config):
@@ -104,6 +100,7 @@ async def run_all_pull_actions(event, gh, config):
 
     await wip_labels.run(event, gh, config)
     await review_labels.run(event, gh, config)
+    await asyncio.sleep(1)
     await wildcard_labels.pending(event, gh)
     await asyncio.sleep(1)
     await wildcard_labels.run(event, gh, config)
@@ -152,7 +149,7 @@ async def run(event, gh, bot):
             cmd = await command_sync(event, gh)
 
         elif m.group('lgtm'):
-            cmd = await command_lgtm(event, m.group('lgtm_key'), gh)
+            cmd = await command_lgtm(event, gh)
 
         else:
             continue